add SSSO collection, rename dsregcmd to azuread
