+++ Seatbelt/Commands/Windows/OSInfoCommand.cs
